requirements.txt
# 多因子分析系统核心依赖
# ==================================

# 数据处理和分析
pandas>=2.0.3
numpy>=1.25.2
scipy>=1.11.1

# 技术指标计算
TA-Lib>=0.4.32
ta>=0.11.0

# HTTP客户端和异步请求
aiohttp>=3.8.4
requests>=2.31.0

# 配置管理
python-dotenv>=1.0.0
PyYAML>=6.0

# 数据可视化（可选）
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.12.0

# 开发和测试工具（可选）
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# 性能监控（可选）
psutil>=5.9.5

# 数据缓存和序列化（可选）
orjson>=3.8.10
